MindCareAgent/requirements.txt
langchain-experimental==0.3.4
langchain-groq==0.2.4
numpy==2.1.3
pandas==2.2.3
pynput==1.7.7
python-dotenv==1.0.1
pyttsx3==2.98
sounddevice==0.5.1